Re-add the TypeTagParser
