+++ ecosystem/typescript/sdk/src/transaction_builder/builder_utils.ts
@@ -45,195 +45,6 @@ function assertType(val: any, types: string[] | string, message?: string) {
   }
 }
 
-function bail(message: string) {
-  throw new Error(message);
-}
-
-function isWhiteSpace(c: string): boolean {
-  if (c.match(/\s/)) {
-    return true;
-  }
-  return false;
-}
-
-function isValidAlphabetic(c: string): boolean {
-  if (c.match(/[_A-Za-z0-9]/g)) {
-    return true;
-  }
-  return false;
-}
-
-type TokenType = string;
-type TokenValue = string;
-type Token = [TokenType, TokenValue];
-
-// Returns Token and Token byte size
-function nextToken(tagStr: string, pos: number): [Token, number] {
-  const c = tagStr[pos];
-  if (c === ":") {
-    if (tagStr.slice(pos, pos + 2) === "::") {
-      return [["COLON", "::"], 2];
-    }
-    bail("Unrecognized token.");
-  } else if (c === "<") {
-    return [["LT", "<"], 1];
-  } else if (c === ">") {
-    return [["GT", ">"], 1];
-  } else if (c === ",") {
-    return [["COMMA", ","], 1];
-  } else if (isWhiteSpace(c)) {
-    let res = "";
-    for (let i = pos; i < tagStr.length; i += 1) {
-      const char = tagStr[i];
-      if (isWhiteSpace(char)) {
-        res = `${res}${char}`;
-      } else {
-        break;
-      }
-    }
-    return [["SPACE", res], res.length];
-  } else if (isValidAlphabetic(c)) {
-    let res = "";
-    for (let i = pos; i < tagStr.length; i += 1) {
-      const char = tagStr[i];
-      if (isValidAlphabetic(char)) {
-        res = `${res}${char}`;
-      } else {
-        break;
-      }
-    }
-    return [["IDENT", res], res.length];
-  }
-  throw new Error("Unrecognized token.");
-}
-
-function tokenize(tagStr: string): Token[] {
-  let pos = 0;
-  const tokens = [];
-  while (pos < tagStr.length) {
-    const [token, size] = nextToken(tagStr, pos);
-    if (token[0] !== "SPACE") {
-      tokens.push(token);
-    }
-    pos += size;
-  }
-  return tokens;
-}
-
-/**
- * Parser to parse a type tag string
- */
-export class TypeTagParser {
-  private readonly tokens: Token[];
-
-  constructor(tagStr: string) {
-    this.tokens = tokenize(tagStr);
-  }
-
-  private consume(targetToken: string) {
-    const token = this.tokens.shift();
-    if (!token || token[1] !== targetToken) {
-      bail("Invalid type tag.");
-    }
-  }
-
-  private parseCommaList(endToken: TokenValue, allowTraillingComma: boolean): TypeTag[] {
-    const res: TypeTag[] = [];
-    if (this.tokens.length <= 0) {
-      bail("Invalid type tag.");
-    }
-
-    while (this.tokens[0][1] !== endToken) {
-      res.push(this.parseTypeTag());
-
-      if (this.tokens.length > 0 && this.tokens[0][1] === endToken) {
-        break;
-      }
-
-      this.consume(",");
-      if (this.tokens.length > 0 && this.tokens[0][1] === endToken && allowTraillingComma) {
-        break;
-      }
-
-      if (this.tokens.length <= 0) {
-        bail("Invalid type tag.");
-      }
-    }
-    return res;
-  }
-
-  parseTypeTag(): TypeTag {
-    if (this.tokens.length === 0) {
-      bail("Invalid type tag.");
-    }
-
-    // Pop left most element out
-    const [tokenTy, tokenVal] = this.tokens.shift()!;
-
-    if (tokenVal === "u8") {
-      return new TypeTagU8();
-    }
-    if (tokenVal === "u16") {
-      return new TypeTagU16();
-    }
-    if (tokenVal === "u32") {
-      return new TypeTagU32();
-    }
-    if (tokenVal === "u64") {
-      return new TypeTagU64();
-    }
-    if (tokenVal === "u128") {
-      return new TypeTagU128();
-    }
-    if (tokenVal === "u256") {
-      return new TypeTagU256();
-    }
-    if (tokenVal === "bool") {
-      return new TypeTagBool();
-    }
-    if (tokenVal === "address") {
-      return new TypeTagAddress();
-    }
-    if (tokenVal === "vector") {
-      this.consume("<");
-      const res = this.parseTypeTag();
-      this.consume(">");
-      return new TypeTagVector(res);
-    }
-    if (tokenTy === "IDENT" && (tokenVal.startsWith("0x") || tokenVal.startsWith("0X"))) {
-      const address = tokenVal;
-      this.consume("::");
-      const [moduleTokenTy, module] = this.tokens.shift()!;
-      if (moduleTokenTy !== "IDENT") {
-        bail("Invalid type tag.");
-      }
-      this.consume("::");
-      const [nameTokenTy, name] = this.tokens.shift()!;
-      if (nameTokenTy !== "IDENT") {
-        bail("Invalid type tag.");
-      }
-
-      let tyTags: TypeTag[] = [];
-      // Check if the struct has ty args
-      if (this.tokens.length > 0 && this.tokens[0][1] === "<") {
-        this.consume("<");
-        tyTags = this.parseCommaList(">", true);
-        this.consume(">");
-      }
-
-      const structTag = new StructTag(
-        AccountAddress.fromHex(address),
-        new Identifier(module),
-        new Identifier(name),
-        tyTags,
-      );
-      return new TypeTagStruct(structTag);
-    }
-
-    throw new Error("Invalid type tag.");
-  }
-}
-
 export function ensureBoolean(val: boolean | string): boolean {
   assertType(val, ["boolean", "string"]);
   if (typeof val === "boolean") {
